requirements.txt
asgiref==3.4.1
Django==3.2.7
django-crispy-forms==1.12.0
django-filter==2.4.0
django-leaflet==0.28.1
django-widget-tweaks==1.4.8
GDAL==3.3.2
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
